Change default path back to shared_path/bundle
The default :bundle_path has always been `shared_path.join('bundle')` and says
so in the README and in the `bundle:install` task description. However during a
recent revert commit, somehow it was changed to
`shared_path.join('vendor/bundle')`. I assume this was done inadvertently.

By fixing this, we avoid a problem when users upgrade capistrano-bundler and
find that their bundler directory has suddenly changed and all their gems are
being re-downloaded.