Separate default :bundle_without value with :
The official bundler documentation specifies that the `without` config
setting should be a colon-separate value. We were using space-separated
for the default value of `"development test"`, which lead to this
warning:

```
You are replacing the current local value of without, which is currently "development:test"
```

Fix by changing the default `:bundle_without` value to be
`"development:test"` (i.e. colon-separated).

Fixes #125